pip_requirements.txt
aiogram            2.22.1
aiohttp            3.8.1
aiosignal          1.2.0
async-timeout      4.0.2
attrs              22.1.0
Babel              2.9.1
certifi            2022.6.15
charset-normalizer 2.1.0
python-dotenv      0.20.0
pytz               2022.2.1
setuptools         60.2.0
wheel              0.37.1
yarl               1.8.1
